# Decoder: use lookup table to validate ASCII